Fixes bug with editing old reports
If a contract that was reported on was now in the state finished,
the edit form would not show it and it would default to the first
contract on the list.

This is now fixed by passing the list of all available contracts,
when editing an old report. Potentially could be restricted by time
or maybe old reports should not be editable at all.

This fixes an issue that caused some confusion in production. So this
solution for now will work, but the flow should be reviewed.