Changes report_parts uniqueness validation to use database index
We were facing an issue when deleting existing report parts and adding
another part with the same contract and role selected. The ActiveRecord
validation occurs before the old records are deleted, so the validation
failed due to duplicates. This doesn't seem to be possible to do at the
moment, so instead we are now using the Database index to catch
duplicates and adding a custom error message.